--- src/app/App.tsx
@@ -1,28 +1,0 @@
-import React from 'react';
-import {MainLayout} from '../components/layout/MainLayout';
-import {AnalyticBoard} from '../features/analyticBoard/components/AnalyticBoard';
-import {ChooseSpecButton} from '../features/chooseSpecFlow/components/ChooseSpecButton';
-import {HeaderLayout} from '../components/layout/HeaderLayout';
-import {MainContentLayout} from '../components/layout/MainContentLayout';
-import {MainMenuItems, MainMenuType} from '../components/VerticalMenu';
-import {Sidebar} from '../features/sidebar/components/Sidebar';
-
-interface Props {
-  activeMenu?: MainMenuType;
-}
-
-const defaultProps: Partial<Props> = {
-  activeMenu: MainMenuItems.VISUALIZATION
-};
-
-export const App = (props: Props) => {
-  return (
-    <MainLayout
-      renderHeader={() => <HeaderLayout renderContent={() => <ChooseSpecButton />} />}
-      renderSideBar={() => <Sidebar activeMenu={props.activeMenu} />}
-      renderContent={() => <MainContentLayout renderCanvasBoard={() => <AnalyticBoard />} />}
-    />
-  );
-};
-
-App.defaultProps = defaultProps;
