+++ src/containers/analyticBoard/AnalyticBoard.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import {VegaLiteTopLevelUnitSpec} from '../../../packages/coreTypes/types';
+import {VegaLiteTopLevelUnitSpec} from '../../packages/coreTypes/types';
 import {useSelector} from 'react-redux';
-import {selectAnalyticBoardMainSpec} from '../../../selectors/analyticBoardSelectors';
-import {AnalyticWidget} from '../../../components/AnalyticWidget';
+import {selectAnalyticBoardMainSpec} from '../../selectors/analyticBoardSelectors';
+import {AnalyticWidget} from '../../components/AnalyticWidget';
 import {EncodingContainer} from './EncodingContainer';
 
 interface Props {
