+++ src/containers/dataSource/DataSourceFieldListContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {DataSourceFieldList} from './DataSourceFieldList';
 import {useSelector} from 'react-redux';
-import {selectDataSourceMetadataAll} from '../../../selectors/dataSourceMetadataSelectors';
+import {selectDataSourceMetadataAll} from '../../selectors/dataSourceMetadataSelectors';
 
 export const DataSourceFieldListContainer = () => {
   const fields = useSelector(selectDataSourceMetadataAll);
